stage2/00-copies-and-fills/02-run.sh: Fix re-runs of script (#301)
Check that file exist before move.

+++ stage2/00-copies-and-fills/02-run.sh
@@ -1,3 +1,6 @@
 ***REMOVED***
 
-mv "${ROOTFS_DIR***REMOVED***/etc/ld.so.preload" "${ROOTFS_DIR***REMOVED***/etc/ld.so.preload.disabled"
+if [ -f "${ROOTFS_DIR***REMOVED***/etc/ld.so.preload" ]; then
+   mv "${ROOTFS_DIR***REMOVED***/etc/ld.so.preload" "${ROOTFS_DIR***REMOVED***/etc/ld.so.preload.disabled"
+fi
+
